refactor: move .test.ts (2)
