+++ app/routes/videos/index.test.ts
@@ -1,10 +1,10 @@
 import { tinyassert } from "@hiogawa/utils";
 import { describe, expect, it } from "vitest";
 import { testLoader, useUser } from "../../misc/test-helper";
-import { loader } from "../bookmarks/index";
+import { loader } from "./index";
 
 // TODO: use data
-describe("bookmarks/index.loader", () => {
+describe("videos/index.loader", () => {
   const { signin } = useUser({
     seed: __filename,
   });
@@ -21,12 +21,7 @@ describe("bookmarks/index.loader", () => {
             "total": 0,
             "totalPage": 0,
           },
-          "request": {
-            "order": "createdAt",
-            "page": 1,
-            "perPage": 20,
-          },
-          "rows": [],
+          "videos": [],
         },
       }
     `);
